# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zoning Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zoning-boundaries:
    get:
      operationId: listZoningBoundaries
      tags: [boundaries]
      summary: List all zoning boundaries
      responses:
        "200":
          description: List of boundaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZoningBoundary"
    post:
      operationId: createZoningBoundary
      tags: [boundaries]
      summary: Create a zoning boundary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoningBoundaryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoningBoundary"
  /zoning-boundaries/{id}:
    patch:
      operationId: updateZoningBoundary
      tags: [boundaries]
      summary: Update a zoning boundary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoningBoundaryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoningBoundary"
    delete:
      operationId: deleteZoningBoundary
      tags: [boundaries]
      summary: Delete a boundary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /kobo/submissions:
    get:
      operationId: listKoboSubmissions
      tags: [kobo]
      summary: List KoboToolbox submissions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KoboSubmission"
  /kobo/sync-logs:
    get:
      operationId: listSyncLogs
      tags: [kobo]
      summary: List sync logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sync logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SyncLog"
  /kobo/config:
    get:
      operationId: getKoboConfig
      tags: [kobo]
      summary: Get KoboToolbox configuration
      responses:
        "200":
          description: Config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KoboConfig"
    patch:
      operationId: updateKoboConfig
      tags: [kobo]
      summary: Update KoboToolbox configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KoboConfigUpdate"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KoboConfig"
  /approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List approval/validation requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: record_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Approvals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Approval"
    post:
      operationId: createApproval
      tags: [approvals]
      summary: Submit a record for approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
  /approvals/{id}/review:
    post:
      operationId: reviewApproval
      tags: [approvals]
      summary: Approve or reject a record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecision"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
...
components:
  schemas:
...
    ZoningBoundary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        zone_type:
          type: string
        geojson:
          type: object
        color:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        created_at:
          type: string
      required: [id, name, zone_type, geojson, created_at]
    ZoningBoundaryInput:
      type: object
      properties:
        name:
          type: string
        zone_type:
          type: string
          enum: [residential, commercial, industrial, agricultural, institutional, protected_area, mixed_use]
        geojson:
          type: object
        color:
          type: string
        description:
          type: string
      required: [name, zone_type, geojson]
    ZoningBoundaryUpdate:
      type: object
      properties:
        name:
          type: string
        zone_type:
          type: string
        geojson:
          type: object
        color:
          type: string
        description:
          type: string
...
    KoboSubmission:
      type: object
      properties:
        id:
          type: integer
        submission_id:
          type: string
        form_id:
          type: string
        data:
          type: object
        synced_at:
          type: string
        processed:
          type: boolean
      required: [id, submission_id, form_id, data, synced_at, processed]
    SyncLog:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [success, failed, in_progress]
        records_synced:
          type: integer
        duplicates_skipped:
          type: integer
        error_message:
          type: ["string", "null"]
        started_at:
          type: string
        completed_at:
          type: ["string", "null"]
      required: [id, status, records_synced, duplicates_skipped, started_at]
    KoboConfig:
      type: object
      properties:
        api_url:
          type: string
        form_id:
          type: string
        is_configured:
          type: boolean
        sync_interval_minutes:
          type: integer
        last_sync:
          type: ["string", "null"]
      required: [api_url, form_id, is_configured, sync_interval_minutes, last_sync]
    KoboConfigUpdate:
      type: object
      properties:
        api_url:
          type: string
        api_token:
          type: string
        form_id:
          type: string
        sync_interval_minutes:
          type: integer
    Approval:
      type: object
      properties:
        id:
          type: integer
        record_id:
          type: integer
        reviewer_id:
          type: ["integer", "null"]
        submitted_by:
          type: integer
        status:
          type: string
          enum: [pending, under_review, approved, rejected]
        comment:
          type: ["string", "null"]
        reviewer_name:
          type: ["string", "null"]
        submitter_name:
          type: string
        record_reference:
          type: string
        created_at:
          type: string
        reviewed_at:
          type: ["string", "null"]
      required:
        - id
        - record_id
        - submitted_by
        - status
        - submitter_name
        - record_reference
        - created_at
    ApprovalInput:
      type: object
      properties:
        record_id:
          type: integer
        submitted_by:
          type: integer
        comment:
          type: string
      required: [record_id, submitted_by]
    ReviewDecision:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected, under_review]
        comment:
          type: string
        reviewer_id:
          type: integer
      required: [status, reviewer_id]
...