# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zoning Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalStatus } from './approvalStatus';

export interface Approval {
  id: number;
  record_id: number;
  /** @nullable */
  reviewer_id?: number | null;
  submitted_by: number;
  status: ApprovalStatus;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  reviewer_name?: string | null;
  submitter_name: string;
  record_reference: string;
  created_at: string;
  /** @nullable */
  reviewed_at?: string | null;
}